+++ near-sdk/src/environment/mod.rs
@@ -1,7 +1,7 @@
-#[cfg(not(target_arch = "wasm32"))]
-pub mod blockchain_interface;
 pub mod env;
 pub mod sys;
 
 #[cfg(not(target_arch = "wasm32"))]
-pub mod mocked_blockchain;
+/// Mock blockchain utilities. These can only be used inside tests and are not available for
+/// a wasm32 target.
+pub mod mock;