+++ near-sdk/src/test_utils/mod.rs
@@ -1,5 +1,3 @@
-use crate::env;
-
 #[allow(dead_code)]
 pub mod test_env;
 
@@ -8,9 +6,8 @@ pub use context::{accounts, testing_env_with_promise_results, VMContextBuilder};
 use near_vm_logic::mocks::mock_external::Receipt;
 
 /// Initializes a testing environment to mock interactions which would otherwise go through a
-/// validator node. This macro will initialize or overwrite the [`BLOCKCHAIN_INTERFACE`]
-/// instance which satisfies the [`BlockchainInterface`] trait for interactions from a
-/// smart contract.
+/// validator node. This macro will initialize or overwrite the [`MockedBlockchain`]
+/// instance for interactions from a smart contract.
 ///
 /// There are five parameters that can be accepted to configure the interface with a
 /// [`MockedBlockchain`], in this order:
@@ -53,9 +50,7 @@ use near_vm_logic::mocks::mock_external::Receipt;
 /// # }
 /// ```
 ///
-/// [`BLOCKCHAIN_INTERFACE`]: crate::env::BLOCKCHAIN_INTERFACE
-/// [`BlockchainInterface`]: crate::BlockchainInterface
-/// [`MockedBlockchain`]: crate::MockedBlockchain
+/// [`MockedBlockchain`]: crate::mock::MockedBlockchain
 /// [`VMContext`]: crate::VMContext
 /// [`VMConfig`]: crate::VMConfig
 /// [`RuntimeFeesConfig`]: crate::RuntimeFeesConfig
@@ -66,18 +61,15 @@ use near_vm_logic::mocks::mock_external::Receipt;
 #[macro_export]
 macro_rules! testing_env {
     ($context:expr, $config:expr, $fee_config:expr, $validators:expr, $promise_results:expr $(,)?) => {
-        $crate::env::set_blockchain_interface(Box::new($crate::MockedBlockchain::new(
+        $crate::env::set_blockchain_interface($crate::MockedBlockchain::new(
             $context,
             $config,
             $fee_config,
             $promise_results,
-            match $crate::env::take_blockchain_interface() {
-                Some(mut bi) => bi.as_mut_mocked_blockchain().unwrap().take_storage(),
-                None => Default::default(),
-            },
+            $crate::mock::with_mocked_blockchain(|b| b.take_storage()),
             $validators,
             None,
-        )));
+        ));
     };
     ($context:expr, $config:expr, $fee_config:expr, $validators:expr $(,)?) => {
         $crate::testing_env!($context, $config, $fee_config, $validators, Default::default());
@@ -94,28 +86,13 @@ macro_rules! testing_env {
 #[allow(dead_code)]
 /// Returns a copy of logs from VMLogic. Only available in unit tests.
 pub fn get_logs() -> Vec<String> {
-    let blockchain_interface =
-        env::take_blockchain_interface().expect("Blockchain interface is not set");
-    let logs = blockchain_interface
-        .as_mocked_blockchain()
-        .expect("MockedBlockchain interface expected")
-        .logs();
-    env::set_blockchain_interface(blockchain_interface);
-    logs
+    crate::env::BLOCKCHAIN_INTERFACE.with(|b| b.borrow().logs())
 }
 
 /// Accessing receipts created by the contract. Only available in unit tests.
 #[allow(dead_code)]
 pub fn get_created_receipts() -> Vec<Receipt> {
-    let blockchain_interface =
-        env::take_blockchain_interface().expect("Blockchain interface is not set");
-    let receipts = blockchain_interface
-        .as_mocked_blockchain()
-        .expect("MockedBlockchain interface expected")
-        .created_receipts()
-        .clone();
-    env::set_blockchain_interface(blockchain_interface);
-    receipts
+    crate::env::BLOCKCHAIN_INTERFACE.with(|b| b.borrow().created_receipts().clone())
 }
 
 /// Objects stored on the trie directly should have identifiers. If identifier is not provided